added 18n for "add" and "no comments", split out topic component from messagecenter
